+++ include/dense_partitioned_phf.hpp
@@ -34,7 +34,7 @@ struct dense_partitioned_phf  //
     // }
 
     template <typename Builder>
-    double build(Builder& builder, build_configuration const& config)  //
+    uint64_t build(Builder& builder, build_configuration const& config)  //
     {
         auto start = clock_type::now();
